Separate interctc logic into a mixin
