+++ nemo/collections/asr/models/ctc_models.py
@@ -28,7 +28,7 @@ from nemo.collections.asr.data.audio_to_text_dali import AudioToCharDALIDataset,
 from nemo.collections.asr.losses.ctc import CTCLoss
 from nemo.collections.asr.metrics.wer import WER, CTCDecoding, CTCDecodingConfig
 from nemo.collections.asr.models.asr_model import ASRModel, ExportableEncDecModel
-from nemo.collections.asr.parts.mixins import ASRModuleMixin
+from nemo.collections.asr.parts.mixins import ASRModuleMixin, InterCTCModelMixin
 from nemo.collections.asr.parts.utils.audio_utils import ChannelSelectorType
 from nemo.core.classes.common import PretrainedModelInfo, typecheck
 from nemo.core.classes.mixins import AccessMixin
@@ -38,7 +38,7 @@ from nemo.utils import logging
 __all__ = ['EncDecCTCModel']
 
 
-class EncDecCTCModel(ASRModel, ExportableEncDecModel, ASRModuleMixin):
+class EncDecCTCModel(ASRModel, ExportableEncDecModel, ASRModuleMixin, InterCTCModelMixin):
     """Base class for encoder decoder CTC-based models."""
 
     def __init__(self, cfg: DictConfig, trainer: Trainer = None):
@@ -116,7 +116,8 @@ class EncDecCTCModel(ASRModel, ExportableEncDecModel, ASRModuleMixin):
                 "E.g., if intermediate_loss_weights = [0.1, 0.3], regular "
                 "loss will have weight of 0.6"
             )
-        if len(self.intermediate_loss_weights) > 0 and not hasattr(self.encoder, 'capture_output_at_layers'):
+        self.interctc_enabled = len(self.intermediate_loss_weights) > 0
+        if self.interctc_enabled and not hasattr(self.encoder, 'capture_output_at_layers'):
             raise ValueError('To use intermediate CTC loss, encoder has to define "capture_output_at_layers" property')
 
     @torch.no_grad()
@@ -544,73 +545,6 @@ class EncDecCTCModel(ASRModel, ExportableEncDecModel, ASRModuleMixin):
             greedy_predictions,
         )
 
-    def add_interctc_losses(
-        self,
-        loss_value: torch.tensor,
-        transcript: torch.tensor,
-        transcript_len: torch.tensor,
-        compute_wer: bool,
-        log_wer_num_denom: bool = False,
-        log_prefix: str = "",
-    ) -> Tuple[torch.Tensor, Dict]:
-        """Adding interCTC losses if required."""
-        if not self.intermediate_loss_weights or not AccessMixin.is_access_enabled():
-            return loss_value, {}
-        self.log()
-        metrics = {f"{log_prefix}final_ctc_loss": loss_value}
-        # if intermediate_loss_weights was set, the encoder has to register
-        # layer_output_X and layer_length_X tensors. We need to apply decoder
-        # to each of them and compute CTC loss.
-        module_registry = AccessMixin.get_module_registry(self.encoder)['']  # key for encoder
-        captured_tensors = []
-        for layer_idx in self.encoder.capture_output_at_layers:
-            try:
-                layer_outputs = module_registry[f"layer_output_{layer_idx}"]
-                layer_lengths = module_registry[f"layer_length_{layer_idx}"]
-            except KeyError:
-                raise RuntimeError(
-                    f"Intermediate layer {layer_idx} was not captured! "
-                    "Check if length of model.encoder.captured_layer_outputs matches "
-                    "length of model.intermediate_loss_weights properties."
-                )
-            if len(layer_outputs) > 1 or len(layer_lengths) > 1:
-                raise RuntimeError(
-                    "Make sure encoder.forward is called exactly one time before interCTC loss is computed."
-                )
-            captured_tensors.append((self.decoder(encoder_output=layer_outputs[0]), layer_lengths[0]))
-
-        for layer_idx, intermediate_result, loss_weight in enumerate(
-            zip(self.encoder.capture_output_at_layers, captured_tensors, self.intermediate_loss_weights)
-        ):
-            inter_loss_value = self.loss(
-                log_probs=intermediate_result[0],
-                targets=transcript,
-                target_lengths=transcript_len,
-                input_lengths=intermediate_result[1],
-            )
-            metrics[f"{log_prefix}inter_ctc_loss_l{layer_idx}"] = loss_value.detach()
-            loss_value += inter_loss_value * loss_weight
-            if compute_wer:
-                self._wer.update(
-                    predictions=intermediate_result[0],
-                    targets=transcript,
-                    target_lengths=transcript_len,
-                    predictions_lengths=intermediate_result[1],
-                )
-                wer, wer_num, wer_denom = self._wer.compute()
-                self._wer.reset()
-                metrics.update({f'{log_prefix}inter_wer_l{layer_idx}': wer})
-                if log_wer_num_denom:
-                    metrics.update(
-                        {
-                            f'{log_prefix}inter_wer_num_l{layer_idx}': wer_num,
-                            f'{log_prefix}inter_wer_denom_l{layer_idx}': wer_denom,
-                        }
-                    )
-
-        # return total loss
-        return loss_value, metrics
-
     # PTL-specific methods
     def training_step(self, batch, batch_nb):
         # Reset access registry
@@ -640,7 +574,7 @@ class EncDecCTCModel(ASRModel, ExportableEncDecModel, ASRModuleMixin):
         # Add auxiliary losses, if registered
         loss_value = self.add_auxiliary_losses(loss_value)
         loss_value, tensorboard_logs = self.add_interctc_losses(
-            loss_value, transcript, transcript_len, compute_wer=(batch_nb + 1) % log_every_n_steps == 0
+            loss_value, transcript, transcript_len, compute_wer=((batch_nb + 1) % log_every_n_steps == 0)
         )
 
         # potentially adding intermediate CTC losses/wers
@@ -722,27 +656,17 @@ class EncDecCTCModel(ASRModel, ExportableEncDecModel, ASRModuleMixin):
 
     def multi_validation_epoch_end(self, outputs, dataloader_idx: int = 0):
         metrics = super().multi_validation_epoch_end(outputs, dataloader_idx)
-        for idx in range(len(self.intermediate_decoding_results)):
-            loss = torch.stack([x[f"val_inter_ctc_loss{idx}"] for x in outputs]).mean()
-            wer_num = torch.stack([x[f"val_inter_wer_num{idx}"] for x in outputs]).sum()
-            wer_denom = torch.stack([x[f"val_inter_wer_denom{idx}"] for x in outputs]).sum()
-            metrics["log"].update(
-                {f"val_inter_ctc_loss{idx}": loss, f"val_inter_wer{idx}": wer_num / wer_denom,}
-            )
-        # in case above loop is not empty, need to also update final loss
-        if self.intermediate_decoding_results:
-            metrics["log"]["val_final_ctc_loss"] = torch.stack([x["val_final_ctc_loss"] for x in outputs]).mean()
+        self.finalize_interctc_metrics(metrics, outputs, prefix="val_")
+        return metrics
 
+    def multi_test_epoch_end(self, outputs, dataloader_idx: int = 0):
+        metrics = super().multi_test_epoch_end(outputs, dataloader_idx)
+        self.finalize_interctc_metrics(metrics, outputs, prefix="test_")
         return metrics
 
     def test_step(self, batch, batch_idx, dataloader_idx=0):
         logs = self.validation_step(batch, batch_idx, dataloader_idx=dataloader_idx)
-        test_logs = {
-            'test_loss': logs['val_loss'],
-            'test_wer_num': logs['val_wer_num'],
-            'test_wer_denom': logs['val_wer_denom'],
-            'test_wer': logs['val_wer'],
-        }
+        test_logs = {name.replace("val_", "test_"): value for name, value in logs.items()}
         return test_logs
 
     def test_dataloader(self):
