+++ nemo/collections/asr/parts/mixins/__init__.py
@@ -13,6 +13,7 @@
 # limitations under the License.
 
 from nemo.collections.asr.parts.mixins.asr_adapter_mixins import ASRAdapterModelMixin
+from nemo.collections.asr.parts.mixins.interctc_mixin import InterCTCModelMixin
 from nemo.collections.asr.parts.mixins.mixins import (
     ASRAdapterModelMixin,
     ASRBPEMixin,
